+++ changelog.md
@@ -1,2 +1,4 @@
 # V2.1.1
-- [packmodules.cpp/simulator.cpp] CHANGED: `StateTransFcn`.
+- [packmodules.cpp] `StateTransFcn`删除$D$矩阵.
+- [simulator.cpp] ADDED: `FLAG_REDUNDANT`.
+- [matmodules.cpp/hpp] ADDED: `MOutput`.